% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.data.R
\name{clean.data}
\alias{clean.data}
\title{This function cleans data and output statistical summary of the dataset
Transfrom the data into a datafrane
Indexes the Coubt column
Replaces NA with 0
Transform the value into a dataframe
Gives a summary of the Counts}
\usage{
clean.data(x)
}
\description{
This function cleans data and output statistical summary of the dataset
Transfrom the data into a datafrane
Indexes the Coubt column
Replaces NA with 0
Transform the value into a dataframe
Gives a summary of the Counts
}
