% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_density.R
\name{usa_density}
\alias{usa_density}
\title{This is a function that will produce a density filled density plot
First it converts the data into a dataframe
Subsets and Transform the Counts into numbers
Creates a filled density plot and removes the NA's}
\usage{
usa_density(x)
}
\description{
This is a function that will produce a density filled density plot
First it converts the data into a dataframe
Subsets and Transform the Counts into numbers
Creates a filled density plot and removes the NA's
}
