% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_hist.R
\name{usa_hist}
\alias{usa_hist}
\title{This function creates transform the data and creates a histogram plot with density curves
First it converts the data into a dataframe
Subsets and Transform the Counts into numbers
Then generates a historgram plot with a density curve}
\usage{
usa_hist(x)
}
\description{
This function creates transform the data and creates a histogram plot with density curves
First it converts the data into a dataframe
Subsets and Transform the Counts into numbers
Then generates a historgram plot with a density curve
}
